Reorganise
Create a new folder for audio and move all the audio specific functions
into it. This folder will contain the necessary functions for the new
AudioPlayer class.
