+++ client/src/components/game/mechanics/audio/makePacmanDeathAudio.js
